apps/edge/requirements.txt
# Edge Controller Dependencies
# Core
asyncio-mqtt==0.16.2
pymodbus==3.6.6
aiosqlite==0.20.0
pyyaml==6.0.2
pydantic==2.6.1
pydantic-settings==2.2.1

# ONNX Runtime for local ML inference
onnxruntime==1.17.1
numpy==1.26.4

# Prometheus metrics
prometheus-client==0.20.0
psutil==5.9.8

# Logging
structlog==24.1.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-mock==3.12.0

# BMS Simulator (dev only)
pymodbus[serial]==3.6.6
